mobile/android/tests/browser/chrome/test_identity_mode.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1099088
-->
<head>
  <meta charset="utf-8">
  <title>Test for getIdentityMode</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript">

  "use strict";

  ChromeUtils.import("resource://gre/modules/Services.jsm");

  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let IdentityHandler = chromeWin.IdentityHandler;

  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:firefox")) == IdentityHandler.IDENTITY_MODE_CHROMEUI,
     "'about:firefox' is a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:config")) == IdentityHandler.IDENTITY_MODE_CHROMEUI,
     "'about:config' is a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:accounts")) == IdentityHandler.IDENTITY_MODE_CHROMEUI,
     "'about:accounts is a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:addonss")) == IdentityHandler.IDENTITY_MODE_UNKNOWN,
     "'about:addonss is not a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:accountss")) == IdentityHandler.IDENTITY_MODE_UNKNOWN,
     "'about:accountss is not a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:accounts?action=signup")) == IdentityHandler.IDENTITY_MODE_CHROMEUI,
     "'about:accounts?action=signup is a verified internal page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("about:evil_extension_page")) == IdentityHandler.IDENTITY_MODE_UNKNOWN,
     "'about:evil_extension_page' is not a verified internal page");

  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("http://mozilla.com")) == IdentityHandler.IDENTITY_MODE_UNKNOWN,
     "http://mozilla.com is an unknown page");
  ok(IdentityHandler.getIdentityMode(0, Services.io.newURI("https://mozilla.com")) == IdentityHandler.IDENTITY_MODE_UNKNOWN,
     "https://mozilla.com over an insecure connection is an unknown page");
  ok(IdentityHandler.getIdentityMode(Ci.nsIWebProgressListener.STATE_IS_SECURE, Services.io.newURI("https://mozilla.com")) == IdentityHandler.IDENTITY_MODE_IDENTIFIED,
     "https://mozilla.com over a secure connection is a verified page");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1099088">Mozilla Bug 1099088</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
